Use the environment clang version for bazel builds
Since we're using environment variables for the CMake build, it's
simpler and less error-prone to use the same clang version everywhere
instead of manually specifying it in the .bazelrc.

Since GitHub doesn't propagate environment variables between steps
(https://github.com/orgs/community/discussions/26407), we need to set
it manually for the entire `build-and-test` job.

PiperOrigin-RevId: 759227069