Drop edge weights along paths with missing path predecessor block [CL 2/2].
Path profiles are intermittent. Therefore, after applying the clonings, we may leave some residual weight in the edges when we didn't have the full path information for them. Specifically, these are edges in the beginning of the LBR path with missing path predecessor blocks.

Our solution implemented here is to drop the edge weights associated with paths with missing predecessor blocks. These paths originate from a node along the cloned path and continue along the path. The edges to drop are the outgoing edge weights (inter and intra) from the last block of the path.

For example, say a path 1->2->3->4 is to be cloned (1 is the path predecessor block). We have profiled paths starting with 2->3->6->... and 2->5->...
Since we don't know if the the predecessor block for these paths was 1 or another block, we fully drop them from the profile. This means we drop the edge weights of (2->3) and (3->6) associated with the first group of paths and edge weight of (2->5) associated with the second group.

This feature allows us to implement a cloning score formula (in subsequent CLs) which rewards converting an entire branch to be not-taken. We have seen about 0.4% improvement in internal benchmarks from this change.

PiperOrigin-RevId: 762507246